SocialMediaapplication/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.63.6
django-extensions==3.2.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
idna==3.7
pillow==10.4.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
